fix: normalization

+++ src/backend/Model/Record.cs
@@ -1,28 +1,28 @@
 
 using System.Text.Json.Serialization;
 public record Transaction(
-    float Amount,
+    double Amount,
     int Installments,
     [property: JsonPropertyName("requested_at")] string RequestedAt);
 
 public record Customer(
-    [property: JsonPropertyName("avg_amount")] float AvgAmount,
+    [property: JsonPropertyName("avg_amount")] double AvgAmount,
     [property: JsonPropertyName("tx_count_24h")] int TxCount24h,
     [property: JsonPropertyName("known_merchants")] string[] KnownMerchants);
 
 public record Merchant(
     string Id,
     string Mcc,
-    [property: JsonPropertyName("avg_amount")] float AvgAmount);
+    [property: JsonPropertyName("avg_amount")] double AvgAmount);
 
 public record Terminal(
     [property: JsonPropertyName("is_online")] bool IsOnline,
     [property: JsonPropertyName("card_present")] bool CardPresent,
-    [property: JsonPropertyName("km_from_home")] float KmFromHome);
+    [property: JsonPropertyName("km_from_home")] double KmFromHome);
 
 public record LastTransaction(
     string Timestamp,
-    [property: JsonPropertyName("km_from_current")] float KmFromCurrent);
+    [property: JsonPropertyName("km_from_current")] double KmFromCurrent);
 
 public record FraudRequest(
     string Id,
@@ -34,20 +34,20 @@ public record FraudRequest(
 
 public record FraudResponse(
     bool Approved,
-    [property: JsonPropertyName("fraud_score")] float FraudScore);
+    [property: JsonPropertyName("fraud_score")] double FraudScore);
 
 public record Reference(
-    [property: JsonPropertyName("vector")] float[] Vector,
+    [property: JsonPropertyName("vector")] double[] Vector,
     [property: JsonPropertyName("label")] string Label);
 
 public record NormalizationConfig(
-    [property: JsonPropertyName("max_amount")] float MaxAmount,
-    [property: JsonPropertyName("max_installments")] float MaxInstallments,
-    [property: JsonPropertyName("amount_vs_avg_ratio")] float AmountVsAvgRatio,
-    [property: JsonPropertyName("max_minutes")] float MaxMinutes,
-    [property: JsonPropertyName("max_km")] float MaxKm,
-    [property: JsonPropertyName("max_tx_count_24h")] float MaxTxCount24h,
-    [property: JsonPropertyName("max_merchant_avg_amount")] float MaxMerchantAvgAmount);
+    [property: JsonPropertyName("max_amount")] double MaxAmount,
+    [property: JsonPropertyName("max_installments")] double MaxInstallments,
+    [property: JsonPropertyName("amount_vs_avg_ratio")] double AmountVsAvgRatio,
+    [property: JsonPropertyName("max_minutes")] double MaxMinutes,
+    [property: JsonPropertyName("max_km")] double MaxKm,
+    [property: JsonPropertyName("max_tx_count_24h")] double MaxTxCount24h,
+    [property: JsonPropertyName("max_merchant_avg_amount")] double MaxMerchantAvgAmount);
 
 public record MccRiskConfig(Dictionary<string, float> Values);
 
@@ -61,7 +61,7 @@ public record FaissResponse(int FraudCount);
 [JsonSerializable(typeof(FaissResponse))]
 [JsonSerializable(typeof(NormalizationConfig))]
 [JsonSerializable(typeof(MccRiskConfig))]
-[JsonSerializable(typeof(Dictionary<string, float>))]
+[JsonSerializable(typeof(Dictionary<string, double>))]
 [JsonSerializable(typeof(string))]
 internal partial class JsonContext : JsonSerializerContext
 {
